Add docs for  new .NET 8 with new render modes, SDK parameters compon… (#333)
## Target
<!--
  Why are you making this change?
 -->
https://github.com/CropperBlazor/Cropper.Blazor/issues/332
Add docs for new .NET 8 with new render modes, SDK parameters components
are not updated timely when using 'blazor.web.js'.
#### Open Questions
<!-- OPTIONAL
- [ ] Use the GitHub checklists to spark discussion on issues that may
arise from your approach. Please tick the box and explain your answer.
-->

## Checklist
<!--
It serves as a gentle reminder for common tasks. Confirm it's done and
check everything that applies.
-->
- [x] Documentation updated
- [ ] Tests cover new or modified code
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] New dependencies added
- [ ] Includes breaking changes
- [ ] Version bumped

## Visuals
<!-- OPTIONAL
Show results both before and after this change. When the output changes,
it can be a screenshot of a trace, metric, or log illustrating the
change.
-->

---------

Co-authored-by: George Radchuk <38187349+ColdForeign@users.noreply.github.com>