Feature/update nuget 1 4 2 (#413)
## Target
<!--
  Why are you making this change?
 -->
Overall, this will reduce the number of iterations with incomplete data
and speed up the transfer when interception the image chunks, in
addition, it gives more opportunities to control the size of the stream
#### Open Questions
<!-- OPTIONAL
- [ ] Use the GitHub checklists to spark discussion on issues that may
arise from your approach. Please tick the box and explain your answer.
-->

## Checklist
<!--
It serves as a gentle reminder for common tasks. Confirm it's done and
check everything that applies.
-->
- [ ] Tests cover new or modified code
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the site documentation
- [ ] I have made corresponding changes to the README, NuGet README file
- [ ] My changes generate no new warnings
- [ ] New dependencies added or updated
- [ ] Includes breaking changes
- [ ] Version bumped

## Visuals
<!-- OPTIONAL
Show results both before and after this change. When the output changes,
it can be a screenshot of a trace, metric, or log illustrating the
change.
-->

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: ColdForeign <tankon46555@gmail.com>
Co-authored-by: George Radchuk <38187349+ColdForeign@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>