Refactor cropper comp, update NuGet packages, fix destroy cropper and js module (#433)
## Target
<!--
  Why are you making this change?
 -->
Refactor cropper comp, update NuGet packages, fix destroy cropper and js
module
#### Open Questions
<!-- OPTIONAL
- [ ] Use the GitHub checklists to spark discussion on issues that may
arise from your approach. Please tick the box and explain your answer.
-->

## Checklist
<!--
It serves as a gentle reminder for common tasks. Confirm it's done and
check everything that applies.
-->
- [x] Tests cover new or modified code
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the site documentation
- [x] I have made corresponding changes to the README, NuGet README file
- [x] My changes generate no new warnings
- [x] New dependencies added or updated
- [ ] Includes breaking changes
- [x] Version bumped

## Visuals
<!-- OPTIONAL
Show results both before and after this change. When the output changes,
it can be a screenshot of a trace, metric, or log illustrating the
change.
-->