Update NuGet Downloads badge to be a link in README, NuGet README file (#441)
## Target
<!--
  Why are you making this change?
 -->
Update NuGet Downloads badge to be a link in README, NuGet README file
#### Open Questions
<!-- OPTIONAL
- [ ] Use the GitHub checklists to spark discussion on issues that may
arise from your approach. Please tick the box and explain your answer.
-->

## Checklist
<!--
It serves as a gentle reminder for common tasks. Confirm it's done and
check everything that applies.
-->
- [ ] Tests cover new or modified code
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the site documentation
- [x] I have made corresponding changes to the README, NuGet README file
- [ ] My changes generate no new warnings
- [ ] New dependencies added or updated
- [ ] Includes breaking changes
- [ ] Version bumped

## Visuals
<!-- OPTIONAL
Show results both before and after this change. When the output changes,
it can be a screenshot of a trace, metric, or log illustrating the
change.
-->